Moved terrain generator
